refactor(data-delivery): decouple network handlers
