+++ adapters/networkHandlerFactory.js
@@ -1,46 +1,37 @@
+/* eslint-disable global-require */
+/* eslint-disable import/no-dynamic-require */
+const path = require("path");
 const {
   networkHandler: GenericNetworkHandler
 } = require("./networkhandler/genericNetworkHandler");
-const {
-  networkHandler: BrazeNetworkHandler
-} = require("../v0/destinations/braze/util");
-const {
-  networkHandler: MarketoNetworkHandler
-} = require("../v0/destinations/marketo/util");
-const {
-  networkHandler: PardotNetworkHandler
-} = require("../v0/destinations/pardot/util");
-const {
-  networkHandler: GoogleAdwordsRemarketingListNetworkHandler
-} = require("../v0/destinations/google_adwords_remarketing_lists/util");
-const {
-  networkHandler: GoogleAdwordsEnhancedConversions
-} = require("../v0/destinations/google_adwords_enhanced_conversions/util");
-const {
-  networkHandler: GA4NetworkHandler
-} = require("../v0/destinations/ga4/utils");
-const {
-  networkHandler: GoogleAdwordsOfflineConversionsNetworkHandler
-} = require("../v0/destinations/google_adwords_offline_conversions/utils");
 
-const {
-  networkHandler: FbPixelNetworkHandler
-} = require("../v0/destinations/facebook_pixel/network-handler");
+const { SUPPORTED_VERSIONS } = require("../routes/utils/constants");
+const { getIntegrations } = require("../routes/utils");
 
-const handler = {
-  generic: GenericNetworkHandler,
-  braze: BrazeNetworkHandler,
-  marketo: MarketoNetworkHandler,
-  pardot: PardotNetworkHandler,
-  google_adwords_remarketing_lists: GoogleAdwordsRemarketingListNetworkHandler,
-  google_adwords_enhanced_conversions: GoogleAdwordsEnhancedConversions,
-  ga4: GA4NetworkHandler,
-  google_adwords_offline_conversions: GoogleAdwordsOfflineConversionsNetworkHandler,
-  facebook_pixel: FbPixelNetworkHandler
+const handlers = {
+  generic: GenericNetworkHandler
 };
 
+// Dynamically import the network handlers for all
+// the supported destinations
+SUPPORTED_VERSIONS.forEach(version => {
+  const destinations = getIntegrations(
+    path.resolve(__dirname, `../${version}/destinations`)
+  );
+  destinations.forEach(dest => {
+    try {
+      handlers[
+        dest
+      ] = require(`../${version}/destinations/${dest}/networkHandler`).networkHandler;
+    } catch {
+      // Do nothing as exception indicates
+      // network handler is not defined for that destination
+    }
+  });
+});
+
 const getNetworkHandler = type => {
-  const NetworkHandler = handler[type] || handler.generic;
+  const NetworkHandler = handlers[type] || handlers.generic;
   return new NetworkHandler();
 };
